Implementation of $lookup operation
Basic rough sketch of FHIR facade server, and example operation ($lookup). It's a pain having to go from a dynamic language into a statically-typed and so anaemic set of models (e.g. the contortions needed to build Parameters when it should be data) , but it should be straightforward now to add the other basic terminology operations.